Move prefixes to fields
