+++ src/main/java/seedu/address/logic/Messages.java
@@ -4,7 +4,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import seedu.address.logic.util.Prefix;
+import seedu.address.model.person.fields.prefix.Prefix;
 import seedu.address.model.person.Person;
 
 /**
