+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -2,11 +2,11 @@ package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.util.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.util.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.util.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.util.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.util.CliSyntax.PREFIX_TAG;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_EMAIL;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_NAME;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_PHONE;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_TAG;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
 import java.util.Collection;
