+++ src/main/java/seedu/address/logic/util/ArgumentMultimap.java
@@ -8,6 +8,7 @@ import java.util.Optional;
 import java.util.stream.Stream;
 
 import seedu.address.logic.Messages;
+import seedu.address.model.person.fields.prefix.Prefix;
 
 /**
  * Stores mapping of prefixes to their respective arguments.
