+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -2,11 +2,11 @@ package seedu.address.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.util.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.util.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.util.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.util.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.util.CliSyntax.PREFIX_TAG;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_EMAIL;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_NAME;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_PHONE;
+import static seedu.address.model.person.fields.prefix.CliSyntax.PREFIX_TAG;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import java.util.ArrayList;
