+++ src/test/java/seedu/address/logic/util/ArgumentTokenizerTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+import seedu.address.model.person.fields.prefix.Prefix;
 
 public class ArgumentTokenizerTest {
